+++ src/main/java/cart/configuration/WebMvcConfig.java
@@ -1,7 +1,7 @@
-package cart;
+package cart.configuration;
 
 import cart.dao.MemberDao;
-import cart.ui.MemberArgumentResolver;
+import cart.configuration.resolver.MemberArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
